SDCICD-1286: set resourceApplyMode to Upsert
as a part of the on-going release versioning effort, adjust the behavior
of the SelectorSyncSet to allow PKO to adopt the resources on a cluster
and not be overwritten by Hive syncing.

Once Hive applies the resources, it will no longer try to resync the
changes continuously which will allow Package Operator to take ownership
of the resources and manage their lifecycle.

https://issues.redhat.com/browse/SDCICD-1286

https://docs.google.com/document/d/1cdJuVFk5OhSR6ceLKyBgSZjfPmsnWFeqab_EJUbI0t4/edit#heading=h.bi1qeu8b4rop

Signed-off-by: Brady Pratt <bpratt@redhat.com>